Made the UI preetier for AddEvent and RemoveEvent pages, added Mon-Wed-Fri and Tues-Thurs occurences, added a slider for reminder priority and removed priority from events class, and changed style.css so that an event/reminder has a different colour based on its category

+++ src/main/java/com/example/demo/Reminders.java
@@ -13,11 +13,11 @@ public class Reminders {
         this.subject = subject;
     }
 
-    public int getPriorityLevel() {
+    public double getPriorityLevel() {
         return priorityLevel;
     }
 
-    public void setPriorityLevel(int priorityLevel) {
+    public void setPriorityLevel(double priorityLevel) {
         this.priorityLevel = priorityLevel;
     }
 
@@ -46,7 +46,7 @@ public class Reminders {
     /**
      * how important the reminder is TODO
      */
-    private int priorityLevel;
+    private double priorityLevel;
 
     /**
      * School work or other
